refactor: move butane parser to dex package
